requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-dotenv
alembic

## not required 
# CLIENT_ID = os.getenv("AZURE_CLIENT_ID")
# CLIENT_SECRET = os.getenv("AZURE_CLIENT_SECRET")
# TENANT_ID = os.getenv("AZURE_TENANT_ID")
# REDIRECT_URI = os.getenv("AZURE_REDIRECT_URI")
# AUTHORITY = f"https://login.microsoftonline.com/{TENANT_ID}"
# SCOPE = ["Mail.Read", "User.Read"]
